DecoratorOverworldMixin Annotations and slight refactor

+++ src/main/java/diarr/caveuberhaul/mixin/ExplosionMixin.java
@@ -24,7 +24,7 @@ public class ExplosionMixin {
     @Inject(method = "doExplosionB",at = @At("TAIL"))
     private void doExplode(boolean particles, CallbackInfo ci)
     {
-        int destruction = (int) Math.round(this.explosionSize*4f);
+        int destruction = Math.round(this.explosionSize*4f);
         //System.out.println(destruction);
         for(int i=0;i<=destruction*2;i++)
         {
